+++ internal/ecosystem/rubygems/const.go
@@ -1,0 +1,7 @@
+package rubygems
+
+const (
+	EcosystemName = "RubyGems"
+	GemFile       = "Gemfile"
+	GemFileLock   = "Gemfile.lock"
+)