+++ tests/ecosystem/rubygems/gem-lock_test.go
@@ -1,0 +1,14 @@
+package rubygems
+
+import (
+	"github.com/1franck/cvepack/internal/ecosystem/rubygems"
+	"testing"
+)
+
+func Test_BuildProjectFromGemFileLock(t *testing.T) {
+	project := rubygems.NewProjectFromGemFileLock("./_fixtures")
+
+	if len(project.Packages()) != 46 {
+		t.Errorf("Expected project to have 46 packages, got %d", len(project.Packages()))
+	}
+}